sync_rpl/requirements.txt
requests
pandas
psycopg2-binary
django-crontab==0.7.1


